+++ src/rt/sched/cown.h
@@ -9,7 +9,7 @@
 #include "base_noticeboard.h"
 #include "multimessage.h"
 #include "schedulerthread.h"
-#include "status.h"
+#include "backpressure.h"
 
 namespace verona::rt
 {
