+++ src/rt/sched/schedulerthread.h
@@ -2,13 +2,13 @@
 // SPDX-License-Identifier: MIT
 #pragma once
 
+#include "backpressure.h"
 #include "cpu.h"
 #include "ds/hashmap.h"
 #include "ds/mpscq.h"
 #include "object/object.h"
 #include "schedulerstats.h"
 #include "spmcq.h"
-#include "status.h"
 #include "threadpool.h"
 
 #include <snmalloc.h>
